LayoutTests/media/modern-media-controls/airplay-button/airplay-button-on.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>AirplayButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const airplayButton = new AirplayButton({ layoutTraits: LayoutTraits.macOS });
airplayButton.on = true;

document.body.appendChild(airplayButton.element);

scheduler.frameDidFire = function()
{
    shouldBeTrue('airplayButton.element.classList.contains("on")');
    debug(`window.getComputedStyle(airplayButton.element).backgroundColor = ${window.getComputedStyle(airplayButton.element).backgroundColor}`);
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
